move confirmEmail to utils
